Usb Gadget HAL: enable oem functions in particular bootmode
There is new requirement to check the physical sim card.
It need oem functions to get information from subsystems.
The bootmode in provisioning station is "ffbm-01".
Therefore, to enable oem functions in new bootmode is necessary.
In addition, the feature between "ffbm-00" and "ffbm-01"
should be the same, so it need to be enabled on two bootmodes.

Bug: 122281469
Test: verify oem functions

Change-Id: I2cb2a91de781f3d2b0508aa5d1360478aa2a6eab
Signed-off-by: raychi <raychi@google.com>