power.stats HAL: optional "delta" arg to debug
This CL adds a built-in delta tracking/calculation facility in
power.stats HAL's implementation of the standard debug() interface.
Right now, doing
    lshal debug android.hardware.power.stats@1.0::IPowerStats/default
only displays the current values of all supported power stats.  With
this CL in place, that command now looks for the optional "delta"
argument at the end of the line.  When it is present, behavior changes
as follows:
* a timestamp is captured for the current values
* a snapshot  is captured of the current values
* deltas from the previous timestamp and snapshot are displayed
* the very first time "delta" is used after boot, the calculated
  deltas show up as "none"

Note that no behavior is changed when the command is run without the
optional "delta" arg, and interspersing non-"delta" commands with
"delta" commands won't have any impact (the timestamp and snapshot are
not modified by non-"delta" commands).

See bug for example output.

Bug: 124388963
Test: Ran on multiple device types with various power consumption
configurations and verified that the output was both complete and
accurate.

Change-Id: Ib8d8ae7bf0e6000f4934d7b6caec13d3f5b3115b