Add a Pixel-specific implementation of IBootControl HAL.
This is in preparation for supporting IBootControl 1.1 for Virtual A/B.

The "legacy" implementation is designed for devices running kernel
versions 4.14 or lower, which rely on the older hardware/qcom/bootctrl
module and a device-specific GPT library. It is currently identical to
the default HAL for IBootControl 1.0. (The default 1.1 implementation
will not wrap legacy libhardware HALs.)

Bug: 141131816
Test: vts
Change-Id: I6fc1ef3f7b8873ab0ad029baf72082bdd202edef