Implement IBootControl 1.1.
Store the new state for IBootControl 1.1 in the misc partition, using
the reserved section of bootloader_message_ab. The code to read/write
this partition is from the bootloader_message library.

We don't embed libboot_control directly because the 1.0 HAL methods are
implemented using GPT, not misc. Note that the new 1.1 methods are
implemented in a shared file. This is for future implementations of the
HAL that may not rely on the legacy GPT library, but will use the same
misc storage.

Bug: 141131816
Test: vts
Change-Id: I455c202cdd6a21261a2ed1f17dec6efd2ae8116e