init.insmod.sh: support modprobe with a * matching pattern
If the module list is specified by a *, or empty, then use contents of
modules.load as argument to modprobe.  Chose this route over enhancing
the modprobe tool to accept an argument of * natively.  Specifically
also support the -b (blacklist) option to limit which modules to load.

Signed-off-by: Mark Salyzyn <salyzyn@google.com>
Bug: 151950334
Test: second stage modules load with * argument in config
Merged-In: Id3b5c43644e465f1327b1a56d11609bab6aad228
Change-Id: Id3b5c43644e465f1327b1a56d11609bab6aad228