Create AndroidTest.xml for pixel vibrator benchmarks
Use tradefed GoogleBenchmarkTest, which is a test
that runs given Google benchmark test package on a device. The results
are parsed to tradefed.metrics.proto.MetricMeasurement.Metric, which
might be available to 'https://android-master-ingest.skia.org/upload'
by using the test-suite-tag 'apct-metric-instrumentation'.

Other examples of benchmark test suites using this (but not tagged):
https://source.corp.google.com/search?q=f:androidtest.xml%20googlebenchmarktest&ct=os&sq=package:android%20-package:aosp

Bug: b/159981541
Test: N/A
Change-Id: I041d8416c81f1fac7bd3c3d1df9daf8e8ab9f567