Add support for Dauntless-equipped phones
We need distinct versions of the init script for different chips,
but we want to keep the same name on the target so we don't have
to change the factory flow and instructions.

But we can't have multiple packages that install the same file
even if they're never used together, so we let the local makefile
handle it.

Bug: 111768050
Test: manual
Change-Id: I741d15cb2e7449bacbc5973e509f16b63ef49f75
Signed-off-by: Bill Richardson <wfrichar@google.com>