Add PixelLogger.mk for Pixelize the makefile setting.
In original design, pixellogger was included in Pixelize mk file,
but the sepolicy are defined by the product specific te file.
These are not aligned and have dependency concern if add new sepolicy rule
in Pixelize te file.
This change put PRODUCT_PACKAGES and BOARD_SEPOLICY into the same file.
It can include all of the necessary config by including this mk file.

Bug: 159650456
Test: Pixel Logger is workable
Change-Id: I768795020578e9d914a8f03b74cd4654b12fdfcb