misc_writer: Write pKVM flag
Communicate the desired pKVM mode to the bootloader by writing to misc.
This configuration will generally be ignored on locked devices but,
otherwise, it can be used to select whether pKVM is enabled or disabled.

This flag will be used in experiments that enable and disable pKVM in
different populations of the dogfood population. It is tristate as
different devices will have pKVM enabled or disabled as their default
state.

Bug: 192819132
Test: toggle pKVM config and observe the change in fastboot
Change-Id: I5fd915e544ed15d99fafd578e525ca347263b8e2