libpixelusb: Do not write into eventfd when temperature is not monitored
b9f33ac34dbe2a1505dcea7f24004c42ea2a8a70 ("thermal: send notification
after callback registered") makes thermal hal to send out a callback,
right after the registration, even when there is no change in the
throttling status. This causes a pending event to be posted to the
eventfd eventhough monitorThread is waiting in the cv_ for the
throttling status to be hot. This prevents system from suspending
as the pending event in evenfd would never be cleared. To prevent
this check for monitorTemperature and bail out if needed.

Bug: 196755565
Test: test for eventfd wakelock upon boot; Test for port overheat.
Signed-off-by: Badhri Jagan Sridharan <badhri@google.com>
Change-Id: I3918e9c41d50948f79f0856c4da62be99df5907d