vibrator: cs40l25: do not check vibe_state
If the request to turn the vibrator on has made it to the HAL then we
can assume that we want to play this new request. Even if there is a
vibration currently playing, we should stop the current vibration and
play the new one. We can assume this because the Framework will handle
all the prioritization of requests before invoking HAL APIs.

Bug: 200067948
Test: request QMC test
Test: typing and receiving a notification
Change-Id: Icfeddc671cf37400212e61798ec422cb880eeb85
Signed-off-by: Vince Leung <leungv@google.com>