Add Adaptive CPU model.
This model was trained on a different device with a different CPU
architecture, so we do not expect it to perform well - we only expect to
test Adaptive CPU end-to-end.

I've #include'd the model code in this CL. This gives us a lot of
separation between the code we're actually touching & the 40kl model
file.

N.B.: We don't currently test the main Adaptive CPU loop, and we're
adding more & more complexity. I propose tackling this in a separate
commit, as it's probably a larger change.

Test: atest libadaptivecpu_test
Bug: 188770301
Change-Id: I00d5c03f836f72e7dbb4c2ea5bdd06cf462a4722