vibrator: cs40l26: fix compose size boundary issue
The data size of the OWT type 10 waveform nsection is just 8 bits. If
the composite array size is 255 with a wait before the first effect,
nsection will become 256 and be truncated as 0 to fit the data size.

The following sequence also make all future OWT upload failed:
1. Normal compose -> 2. Too long compose -> 3. Too long compose
2nd step: Delete the previous memory and failed to upload.
3rd step: Failed to delete and the id is still > -1.
...
Nth step: Failed to delete and the id is still > -1.

Bug: 189301872
Test: atest VtsHalVibratorTargetTest
Signed-off-by: Tai Kuo <taikuo@google.com>
Change-Id: I988c5bc4ca0d18e088be71fae4b8332f9e014acd