Set mIsEnabled=false on error in AdaptiveCpu.
This is instead of exiting the thread. This means we don't have to worry
about restarting the thread, which makes this code a lot simpler (see
ag/16258826).

We also have to move the initialization code into the loop. This is so
that we can re-attempt initializing & prevent doing other actions if
initialization failed. (N.B.: We do not want to do this initialization
in the main thread as it involves IO.)

Test: manual
Bug: 188770301
Change-Id: Ic44d7b6a620a3ff457e3442e3b61da2cc1827c24