Extract work duration processing to a separate class.
Done ahead of adding more features. This won't be a lot of code, but it
will be useful to test it. Extracting this logic to separate class makes
the unit tests a lot easier. Plus, the AdaptiveCpu class is getting
bigger so it's good to extract what we can.

Also, I've decoupled the TakeWorkDurations method from the "wait for
frames" logic. The "wait for frames" logic is fundamental to how ACPU
currently works (if we get disabled/crash, we block receiving frames,
thus disabling ACPU) so it makes sense to separate it from the rest of
the logic.

Test: atest libadaptivecpu_test
Bug: 207366032
Change-Id: I8d53770a4c91fb3d9866d9e818e1bbbea682cb76