Refactor CpuLoadReader interface & introduce ICpuLoadReader.
Done ahead of adding two different implementations of ICpuLoadReader:
one for reading from idlestats (b/208621595), one for reading kernel
modules (b/207662659).

Also refactor the interface, so we don't have an intermediate
representation of CPU loads. This will have a really small gain to
efficiency, but the implementation is a lot simpler this way.

Also move some of the logging code (DumpToStream, logging features) to
inside CpuLoadReader. This allows us to expose less state (which will be
important for other implementations) and reduces the size of the
AdaptiveCpu class.

Test: atest libadaptivecpu_test
Bug: 208621595, 207662659
Change-Id: Id0a988e1868400f1180f316406123d8d525ca44d