libperfmgr: support mask as vote
For a system have following hints:
* Hint A masks hint X, Y;
* Hint B masks hint Y, Z;

Consider a sequence of below:
* Hint A activates - masks X, Y
* Hint B activates - masks Y, Z
* Hint A deactivates - unmasks X, Y

With current implementation, both X and Y are unmasked, Z is masked in
the end.

This CL changes makes libperfmgr remember which hint mask/unmask. So
after this change, in the end, X is unmasked, Y and Z are masked.

Bug: 216063139
Test: boot
Test: /data/nativetest64/libperfmgr_test/libperfmgr_test
Change-Id: I2817d223a04e0b3cc6054efede8ed7e1eae827dd
Signed-off-by: Wei Wang <wvw@google.com>