Add random_throttle_options config to Adaptive CPU.
We'll need this config for collecting traces in the field. These will
need to (rarely) randomly throttle, but if we're training a model to use
a subset of the actions (e.g. just NO_THROTTLE and THROTTLE_70) then we
don't want randomly choose other actions (e.g. THROTTLE_80).

Also, instead of just logging a warning, fail if the random throttle
probabilty >100%.

Test: atest libadaptivecpu_test:AdaptiveCpuConfigTest
Bug: 219743541
Change-Id: Ie51bd5a4641176998a07d5ccea4933fbd50ef6eb