Update model with tag `acpu-2202_EX01_IT02_TR01_plain`.
vvaldes@ has been retraining the model for Pixel 6 Pro. While this won't
be the final model, having a sensible model means we can start running a
coin flip experiment & test our field infra.

As this will take a while to make it to daily/weekly builds, I'd like to
submit this ASAP so we can kick off experiments in 1-2 weeks.

Bug: 192553577
Test: m
Change-Id: Ic15f67eb5c6700b78c4929bdad1880d5de44407f