Prevent hint timeouts when same hint is repeatedly sent.
If we repeatedly send the same hint (as we do in lab test to constantly
throttle) then we will only send the hint once, as we don't send a hint
if the hint doesn't change. The timeout still applies to these hints,
resulting in the hint only lasting 2s, despite being consistently
renewed.

This problem is unlikely to significantly effect ACPU, as we often
change the throttle level more frequently than every 2s. However, this
will still have some impact.

Thanks vvaldes@ for spotting this!

Bug: 230316778
Test: manual
Change-Id: I255cfcfbd732782cfa60f2b86d785ef4d62d3ead