Fix cancelling hints when refreshing expiring hints.
When choose a new throttle option, we:
1. Send the new hints.
2. Cancel the old hints.

After b/230316778, we also sometimes do this when the hint *doesn't
change*, so we can "refresh" the hint if it's about to expire.

This send & immediately cancels the same hints, which essentially just
cancels the hint.

The solution is to only do step #1 from above when refreshing the hint.

Bug: 232782573
Test: manual
Change-Id: I2a6d83a29b683d7e1f8d71b8fff700b51d1839cf