cs40l26: fix memory leaks
`ch` seems to be leaked in many potential error cases in this file,
since the ownership of it is controlled by `Vibrator::on`, which may not
be reached. Since the overload of `on` that takes a `dspmem_chunk*` is
private, we can safely pivot to a story where the previously-mentioned
overload takes no ownership of the given `dspmem_chunk`. This lets us
give ownership to callers, which in turn lets us use either the stack or
`unique_ptr`s to safely keep `depmem_chunk`s.

Also transfer ownership of `dspmem_chunk->head` to `dspmem_chunk`
itself; it looks like `uploadOwtEffect` immediately copies the bytes
into a self-managed buffer. Hence, we need to take responsibility for
freeing that.

Caught by Clang's static analyzer.

Bug: 206470603
Test: TreeHugger
Change-Id: I75763a8c687f83f93e8c7069fd22d0c3d364da06