cs40l26: Prevent waveform number out of range
When we call the cancel vibration right after
a long compose effect, the erase command will timeout and
the active ID was set to -1.

After the effect finished, we will based on the active ID do
the erase function if it's not -1.

So, there is a chance that the active ID is -1 and the compose effect
was not erased, which will also cause an error when testing the VTS.

To fix this, I add this patch to solve this problem.

Bug: 261832151
Test: atest VtsHalVibratorTargetTest
Change-Id: I63a8455962061edce2a9615084aac6e481c91012
Signed-off-by: Chase Wu <chasewu@google.com>