cs40l25: add state polling timeout
Replacing the infinite vibe_state checking to prevent the abnormal non-stop
symptom which would cause HAL stuck on the checking step.

Even if there is a vibration currently playing, we should stop
the current vibration and play the new one.

We can assume this because the Framework will handle
all the prioritization of requests before invoking HAL APIs.

Bug: 250074997
Test: Use previous non-stop FW to test it
Change-Id: I5dfc0cd4eb581d4db23094e9c73ce8e08ad85ea6