Create common utils for AIDL Gadget HAL
Initial GadgetHAL migration to AIDL missed migrating UsbGadgetUtils to
AIDL. Although a UsbGadgetAidlUtils.h was added, it still relied on
HIDL interface for the header's implementation.

This CL separates out the logic common in HIDL and AIDL to CommonUtils
and moves interface specific logic to hidl/ and aidl/ directories. It
also adds libpixelusb-aidl build target for AIDL HALs to depend on.

Note: UsbOverheatEvent still uses HIDL interfaces, so added a TODO for
that.

Bug: 218791946
Test: Check logs to verify that AIDL HALs are using the correct
      implementation
Change-Id: Ie63b62b5e612473993763dfd1640b404bc57414c