Add slow frame detection to PowerHAL
Basically, we just take the last N updates and check if the median
time between those updates is longer than the stale timeout. If so, we
skip the PID logic in reportActualWorkDuration. This approach seems
reasonably robust while still being fairly simple, and all the
parameters are adjustable.

For now, we're using N=4. Larger N value makes it less sensitive to
short-term issues but makes it take longer to kick in. Even values for
N are preferred so that the intervals are odd and the median makes more
sense, since #intervals = N-1. This is not a requirement, however.

Bug: 232329572
Test: manual
Change-Id: I5177c5b16b7f95e0b717c0704d7e700ec249a4e4