misc_writer: add max_ram_size configuration
Communicate the max_ram_size settings with bootloader by writing to misc.
This configuration will be used in experiments that limit the maximum
ddr size on the devices instead of the hardware ddr size.
Bootloader will generally ignore this configuration on secured devices.

Ignore the clang_format and cpplint from repo hooks.

Bug: 279871595
Test: Run misc_writer_test
Test: misc_writer --set-max-ram-size 8192, and check misc partition
Test: misc_writer --set-max-ram-size -1, and check misc partition
Change-Id: I9546e9dceba341a92eff0c769560114bd069760c