thermal: Refactor StatsConfig & Reorder initializtion order.
- Split parseStatsConfig into 2 separate functions for sensor and
  cooling device and remove wrapper struct StatsConfig
To clean up the number of arguments split the function into two one
for sensor and one for cooling devices.
Also as we'll add parsing for abnormal stats, this will help to keep the
functions concise.

Reorder initializtion order to align with declaration order to remove
error "field designators to be specified in declaration order"
- binded_cdev_info_map
- power_status_map_

Bug: 238326225
Test: Build and verify stats logs
Change-Id: If9ba3dbd6abdcf8f881f77a697a34546d7392bdb