libpixelusb: fix free after use in getI2cBusNumberString
getI2cBusNumberString closes the held DIR reference before
making a strtok call to a string that depends on the DIR.
So, cache the result of the strtok call before closing the
directory.

In addition, explicity return the i2c bus in getI2cClientPath
as a string type, and remove the default return for the client
Id. Without the removal, devices that need to test multiple
hsi2c paths will always return some i2c client on the first
path that exists instead of testing all possible paths.

Test: validate usb function on fullmte,
      validate path checking by checking incorrect hsi2c
      addresses before real one on device
Bug: 323771351
Bug: 318820040
Change-Id: I33785b52e23b63b6d60cb3ba68fcebcf9a4e66d9
Signed-off-by: RD Babiera <rdbabiera@google.com>