Adpf: Add heuristic boost to adpf sessions.
Heuristic boost will be enabled if the number of missed cycles is
greater than a threshold. It will be withdrawed when the number of
missed cycles is smaller than another threshold and the reported
duration variance drops below a level.

Bug: 204204746
Test: atest libadpf_test; check related trace on device adpf sessions.
Change-Id: I2aaf47ea09ea7ce37e1648d48f7f7d0050cf70c4
Merged-In: I2aaf47ea09ea7ce37e1648d48f7f7d0050cf70c4