thermal: Fix avg power calculation
Since both energy sample and duration are of type uint64 so when we
subtract them and assign to auto variable, it will also be of type
uint64 which means for negative value cases the value will wrap around
to some large value.

This large value would be not logged and would also cause issue in our
pid loop.

Fix it directly comparing the current and previous energy sample's
duration and counter.

Also update the error log to print the actual timestamp and
energy_counter.

Bug: 326029066
Bug: 325851052

Test: build and verify by passing harcoded values.
Change-Id: I6049698b14b00d19dbb512c078cca8c9c64ffe76