Add implementation for Power Efficiency mode
Added an implementation for the POWER_EFFICIENCY mode, and cleaned
up a bit of existing code in the process. This adds a new flag to
the SessionValueEntry to indicate whether a special min value
defined in the PowerHint.json should be applied.

Bug: 277285195
Test: atest libadpf_test
Change-Id: Id707b6d665aa48721f560b55d8371c2836255109
Merged-In: Id707b6d665aa48721f560b55d8371c2836255109